+++ app/src/main/java/com/azzuresolutions/videocompressor/activity/VideoToAudioActivity.kt
@@ -87,12 +87,12 @@ class VideoToAudioActivity : AppCompatActivity() {
 
             override fun onStartTrackingTouch(seekBar: SeekBar) {
                 mediaPlayer.seekTo(binding.playProgressBar.progress)
-                binding.videoView.seekTo(binding.playProgressBar.progress.toLong())
+                binding.videoView.seekTo(binding.playProgressBar.progress)
             }
 
             override fun onStopTrackingTouch(seekBar: SeekBar) {
                 mediaPlayer.seekTo(binding.playProgressBar.progress)
-                binding.videoView.seekTo(binding.playProgressBar.progress.toLong())
+                binding.videoView.seekTo(binding.playProgressBar.progress)
             }
         })
         binding.imgPlay.setOnClickListener {
